Memory leak detection
Record all allocs inside a allocation size range, and remove on free. Dump
any remaining records on kext unload.

Record stack inside vmem like illumos does. Nothing uses it though.

Alert about size==0 and kmem_free(NULL)

Use the stack-recording instead of macro wrapping - which gives better
location information, and recompile of ZFS not required.

Windows port is unable to resolve function addresses to symbol names,
but debugger can (dt address)

BreakPoint() is left in driver.c to hold at unload so the
cbuf can be saved out.

+++ ZFSin/spl/include/sys/kmem.h
@@ -72,9 +72,9 @@ extern uint64_t physmem;
 #define kmem_zalloc(size, kmflags)  zfs_kmem_zalloc((size), (kmflags))
 #define kmem_free(buf, size)        zfs_kmem_free((buf), (size))
 
-    void* zfs_kmem_alloc(uint32_t size, int kmflags);
-    void* zfs_kmem_zalloc(uint32_t size, int kmflags);
-    void zfs_kmem_free(void *buf, uint32_t size);
+    void* zfs_kmem_alloc(size_t size, int kmflags);
+    void* zfs_kmem_zalloc(size_t size, int kmflags);
+    void zfs_kmem_free(void *buf, size_t size);
 
     void spl_kmem_init(uint64_t);
     void spl_kmem_thread_init();
